+++ test/libsolidity/astPropertyTests/forLoop/isSimpleCounterLoop_negative_cases.sol
@@ -20,6 +20,7 @@ function h() pure returns (uint) {
 
 contract C {
     uint z = 0;
+    uint transient t;
 
     function modifyStateVarZ() public returns (uint) {
         z = type(uint).max;
@@ -117,6 +118,12 @@ contract C {
         for (z = 1; z < 2048; ++z) {
             modifyStateVarZ();
         }
+        /// StorageStateVarCounter: isSimpleCounterLoop
+        for (z = 1; z < 2048; ++z) {
+        }
+        /// TransientStateVarCounter: isSimpleCounterLoop
+        for (t = 1; t < 2048; ++t) {
+        }
         /// NonIntegerCounter: isSimpleCounterLoop
         for (address i = address(0x123); i < address(this); i = address(0x123 + 1)) {
         }
@@ -135,6 +142,8 @@ contract C {
         }
     }
 }
+// ====
+// EVMVersion: >=cancun
 // ----
 // AdditionLoopExpression: false
 // ShortHandAdditionLoopExpression: false
@@ -160,6 +169,8 @@ contract C {
 // LiteralDifferentCommonTypeConditionRHS: false
 // StateVarCounterModifiedFunctionConditionRHS: false
 // StateVarCounterModifiedFunctionLoopBody: false
+// StorageStateVarCounter: false
+// TransientStateVarCounter: false
 // NonIntegerCounter: false
 // AssemblyAfterAssignmentToCounter: false
 // ExpressionIncrement: false