+++ test/libsolidity/astPropertyTests/transient_data_location.sol
@@ -7,6 +7,8 @@ contract C {
     uint y;
 
 }
+// ====
+// EVMVersion: >=cancun
 // ----
 // TransientDataLocation: transient
 // TransientVarName: x