+++ test/libsolidity/semanticTests/operators/transient_storage_variable_increment_decrement.sol
@@ -1,0 +1,16 @@
+contract C {
+    int transient x;
+    function f() public returns (int) {
+        ++x;
+        ++x;
+        --x;
+        x++;
+        x--;
+        return x;
+    }
+}
+// ====
+// EVMVersion: >=cancun
+// compileViaYul: false
+// ----
+// f() -> 1