+++ test/libsolidity/semanticTests/variables/delete_transient_state_variable.sol
@@ -1,0 +1,13 @@
+contract C {
+    uint transient x;
+    function f() public returns (uint) {
+        x = 10;
+        delete x;
+        return x;
+    }
+}
+// ====
+// EVMVersion: >=cancun
+// compileViaYul: false
+// ----
+// f() -> 0