+++ Changelog.md
@@ -4,6 +4,7 @@ Language Features:
 
 
 Compiler Features:
+ * Code Generator: Transient storage value type state variables are now supported by the legacy pipeline.
 
 
 Bugfixes: